Edited the model and organised the code.
